+++ platformio-project/lib/RangeDataActor/RangeDataActor.h
@@ -18,7 +18,7 @@
 /// if inRangeIsOn==false, the output is LOW if the channel is in the range and HIGH if channel value is out of range
 class RangeDataActor : public IDataActor {
        public:
-              RangeDataActor(std::string observedChannel,double lowerBound, double upperBound,uint8_t gpio, bool inRangeIsOn );
+              RangeDataActor(std::string observedChannel,float lowerBound, float upperBound,uint8_t gpio, bool inRangeIsOn );
               void init();
               void reInit();
               void sleep();
@@ -27,12 +27,12 @@ class RangeDataActor : public IDataActor {
               std::string getName();
        protected:
               std::string _observedChannel;
-              double _lowerBound = 0.0f;
-              double _upperBound= 999999999.0f; 
+              float _lowerBound = 0.0f;
+              float _upperBound= 999999999.0f; 
               bool _inRangeIsOn=true;
               uint8_t _gpio = 0;
-              bool isInRange(double value);
-              double extractChannelValue(IDataStorage *dataStorage);
+              bool isInRange(float value);
+              float extractChannelValue(IDataStorage *dataStorage);
 };
 
 #endif
